+++ Glimps/CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2024-11-14 - 1.14.0
+
+### Added
+
+- Waitfor action : submit a file to Glimps detect and wit for analysis to finish
+- Get status action : get Glimps detect profile status
+
+### Changed
+
+- use Glimps py-gdetect client to perform actions
+
 ## 2024-05-28 - 1.13.0
 
 ### Changed